requirements.txt
python-dotenv
aiogram
yt-dlp